Set up Prettier and pin to v2
Initially I wanted to skip any complexity unrelated to Git for this
repo, because formatting of code really doesnt matter in this workshop.
But after issues with Prettier 3 and the default VS Code extension
esbenp.prettier-vscode (that are documented and not yet solved) I tried
disabling the extension for a while only to realise that yeah I really
do want my editor to format on save like I am used to – and I suspect
most participants in the workshop will have en editor set up to do so.

I hope that intentionally setting up Prettier, and also pinning the
version to v2, will create the least risk of formatting hassle for us
so we can focus on practicing Git.

The .prettierrc will instruct my editor that there are formatting rules
to follow on save. Which rules? An empty config file means the default
Prettier config without any overrides.