Set up structure for adventures
This setup will create a collection of adventures and render them as
entrances on the front page. Eleventy will default to sort items by
date, which is often great! but not for this project, so this will also
config eleventy to randomize the collection.

The adventures.json will add front matter for layout and permalink for
all files in the same directory in one place. And yes, I could have
used that single file also for the tags but I will soon want multiple
tags to be able to create different collections. And explicit tags in
will be simpler in the workshop when we mostly work in the .md files.