Change adventures from randomized to alphabetical
I made this change because initially I thought it would be fun to have
the adventures presented randomized. There was no logical order, and I
thought random would lead to cool exploration. But now that I have
added more and more adventures, I started to notice that it has become
annoying to not easily see the specific entrance I am looking for.

Perhaps a future feature could be if they are presented alphabetically
at first, but with a button in the UI to shuffle them. But that is not
happening on the eve before the workshop, so alphabetical only for now.